fix: fail to pre bundle deps which has css
