+++ packages/ice/src/service/serverCompiler.ts
@@ -1,11 +1,10 @@
 import * as path from 'path';
-import { createHash } from 'crypto';
 import consola from 'consola';
 import esbuild from 'esbuild';
 import type { Config } from '@ice/webpack-config/esm/types';
 import lodash from '@ice/bundles/compiled/lodash/index.js';
 import type { TaskConfig } from 'build-scripts';
-import { getCompilerPlugins } from '@ice/webpack-config';
+import { getCompilerPlugins, getCSSModuleLocalIdent } from '@ice/webpack-config';
 import type { ServerCompiler } from '../types/plugin.js';
 import type { UserConfig } from '../types/userConfig.js';
 import escapeLocalIdent from '../utils/escapeLocalIdent.js';
@@ -130,17 +129,8 @@ export function createServerCompiler(options: Options) {
         cssModulesPlugin({
           extract: false,
           generateLocalIdentName: function (name: string, filename: string) {
-            const hash = createHash('md4');
-            hash.update(Buffer.from(filename + name, 'utf8'));
-            const localIdentHash = hash.digest('base64')
-              // Remove all leading digits
-              .replace(/^\d+/, '')
-              // Replace all slashes with underscores (same as in base64url)
-              .replace(/\//g, '_')
-              // Remove everything that is not an alphanumeric or underscore
-              .replace(/[^A-Za-z0-9_]+/g, '')
-              .slice(0, 8);
-            return escapeLocalIdent(`${name}--${localIdentHash}`);
+            // Compatible with webpack css-loader.
+            return escapeLocalIdent(getCSSModuleLocalIdent(filename, name));
           },
         }),
         assetsManifest && createAssetsPlugin(assetsManifest, rootDir),
